fetch userId from session

+++ src/app/api/v1/posts/[post_id]/unlike/route.ts
@@ -1,11 +1,27 @@
 import dbConnect from "@/lib/dbConnect";
 import LikeModel from "@/models/Like.model";
 import { NextRequest, NextResponse } from "next/server";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../../auth/[...nextauth]/options";
 
 export async function DELETE(req:NextRequest,{params}:{params:{post_id:string}},res:NextResponse) {
     await dbConnect();
+
+     //Getting userId from logged in user from Session
+
+     const session = await getServerSession(authOptions)
+     const _user = session?.user;
+ 
+     if(!session || !_user){
+         return NextResponse.json(
+             { success: false, message: 'Not authenticated' },
+             { status: 401 }
+           );
+     }
+ 
+     const userId = _user._id;
         try {
-            const {userId} =await req.json()
+            
             const {post_id}  = params
 
             const unlike = await LikeModel.findOneAndDelete(
